+++ bigtree/tree/export/__init__.py
@@ -1,6 +1,11 @@
 from .dataframes import tree_to_dataframe, tree_to_polars  # noqa
 from .dictionaries import tree_to_dict, tree_to_nested_dict  # noqa
-from .images import tree_to_dot, tree_to_mermaid, tree_to_pillow  # noqa
+from .images import (  # noqa
+    tree_to_dot,
+    tree_to_mermaid,
+    tree_to_pillow,
+    tree_to_pillow_graph,
+)
 from .stdout import (  # noqa
     hprint_tree,
     hyield_tree,
@@ -20,6 +25,7 @@ __all__ = [
     "tree_to_nested_dict",
     "tree_to_dot",
     "tree_to_pillow",
+    "tree_to_pillow_graph",
     "tree_to_mermaid",
     "tree_to_newick",
 ]